requirements.txt
Pillow~=9.3.0
PyQt5~=5.15.7
pydub~=0.25.1